pyproject.toml
[project]
name = "edcat-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firebase-admin>=7.3.0",
    "flask>=3.1.3",
    "flask-babel>=4.0.0",
    "google-cloud-secret-manager>=2.26.0",
    "gunicorn>=25.2.0",
    "python-dotenv>=1.2.2",
]
